+++ lib/filter_deletes.js
@@ -1,6 +1,6 @@
 // TODO: We do not consider forks and joins here: We need to
 
-var errors = require('../lib/errors')
+var errors = require('../errors')
 
 module.exports = function filterSafeDeletes (osm, batch, cb) {
   // A mapping of ids of ways and relations to
