+++ lib/replace_ids.js
@@ -1,5 +1,5 @@
 var util = require('../lib/util')
-var errors = require('../lib/errors')
+var errors = require('../errors')
 
 /**
  * Replace placeholder ids with generated unique ids
