+++ lib/validate_changeset.js
@@ -1,4 +1,4 @@
-var errors = require('../lib/errors')
+var errors = require('../errors')
 
 /**
  * Check:
