+++ routes/changeset_upload.js
@@ -3,7 +3,7 @@ var osm2Obj = require('osm2json')
 var toOsm = require('obj2osm')
 var fromArray = require('from2-array')
 
-var errors = require('../lib/errors')
+var errors = require('../errors')
 var isValidContentType = require('../lib/util').isValidContentType
 
 module.exports = function (req, res, api, params, next) {
