+++ routes/element_multi_fetch.js
@@ -2,7 +2,7 @@ var qs = require('query-string')
 var toOsm = require('obj2osm')
 var fromArray = require('from2-array')
 
-var errors = require('../lib/errors')
+var errors = require('../errors')
 var cmpFork = require('../lib/util').cmpFork
 
 module.exports = function (req, res, api, params, next) {
