+++ routes/misc_capabilities.js
